Fix Ray executor futures to resolve asynchronously
